Rename EndPhase to Epilogue and proceed() to perform()
